fix: impossible to use directly in browser
